MP-1136 Upgrade numpy (#14)
I want to upgrade scipy and scikit-learn in other repos, so that I could run tests for Python 3.11, and for this I want to make sure, that all environments have the same numpy version. So I'm copying the newer version from #12.

This should also eventually fix the warning about mismatching numpy versions between client, scheduler and workers.

In addition I am fixing broken pytorch builds: our current pytorch version is not compatible with the latest setuptools, but it does not set an upper boundary.